♻️refactor/header 컴포넌트 분리 및 props 타입 지정
