+++ src/header/DetailStationSearch.tsx
@@ -1,11 +1,9 @@
-import React, { useEffect, useState } from 'react';
-import { FaSearch } from 'react-icons/fa';
-import { useRecoilState, useRecoilValue } from 'recoil';
-import {
-  currentStationAtom,
-  stationNameAtom,
-  isDropdownAtom,
-} from '../recoil/atoms';
+import { useEffect, useState } from 'react';
+
+import { useRecoilValue } from 'recoil';
+import { currentStationAtom, stationNameAtom } from '../recoil/atoms';
+import DetailStationSearchDropdown from './DetailStationSearchDropdown';
+import DetailStationSearchInput from './DetailStationSearchInput';
 
 function DetailStationSearch() {
   const [searchInput, setSearchInput] = useState('');
@@ -13,8 +11,7 @@ function DetailStationSearch() {
   const [filteredStationNames, setFilteredStationNames] = useState<string[]>(
     [],
   );
-  const [currentStation, setCurrentStation] = useRecoilState(currentStationAtom);
-  const [isDropdown, setIsDropdown] = useRecoilState(isDropdownAtom);
+  const currentStation = useRecoilValue(currentStationAtom);
   const stationNameDatas = useRecoilValue(stationNameAtom);
 
   useEffect(() => {
@@ -42,76 +39,18 @@ function DetailStationSearch() {
     stationNameDatas.line4,
     stationNameDatas.line5,
   ]);
-  const searchInputChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchInput(() => e.target.value);
-    if (e.target.value.replace(/(\s*)/g, '') === '') {
-      setFilteredStationNames(stationNames);
-    } else if (e.target.value.replace(/(\s*)/g, '') !== '') {
-      setFilteredStationNames(
-        stationNames.filter((stationName) =>
-          stationName.includes(e.target.value),
-        ),
-      );
-    }
-  };
-
-  const searchClickHandler = () => {
-    setCurrentStation(searchInput.replace(/(\s*)/g, ''));
-  };
-
-  const searchDropdownClickHandler = (stationName: string) => {
-    setCurrentStation(stationName);
-    setIsDropdown(false);
-  };
-
-  const inputOnFocusHandler = () => {
-    setIsDropdown(true);
-  };
-
-  const dropdownCloseHandler = () => {
-    setIsDropdown(false);
-  };
 
   return (
     <div>
-      {isDropdown ? (
-        <div className="DetailStationSearchDropdown">
-          <div className="DetailStationSerchDropdownSpace" />
-          <div className="DetailStationSearchDropdownWapper">
-            {filteredStationNames.map((filteredStationName) => (
-              <div
-                role="presentation"
-                onClick={() => {
-                  searchDropdownClickHandler(filteredStationName);
-                }}
-              >
-                {filteredStationName}
-              </div>
-            ))}
-          </div>
-          <button
-            className="DetailStationSearchDropdownButton"
-            type="button"
-            onClick={dropdownCloseHandler}
-          >
-            x
-          </button>
-        </div>
-      ) : null}
-      <div className="DetailStationSearch">
-        <input
-          className="DetailStationSearchInput"
-          placeholder="역명 검색"
-          onChange={searchInputChangeHandler}
-          value={searchInput}
-          onFocus={inputOnFocusHandler}
-        />
-        <FaSearch
-          className="DetailStationSearchIcon"
-          size="24"
-          onClick={searchClickHandler}
-        />
-      </div>
+      <DetailStationSearchDropdown
+        filteredStationNames={filteredStationNames}
+      />
+      <DetailStationSearchInput
+        searchInput={searchInput}
+        stationNames={stationNames}
+        setSearchInput={setSearchInput}
+        setFilteredStationNames={setFilteredStationNames}
+      />
     </div>
   );
 }
