+++ src/interfaces/interfaces.ts
@@ -1,3 +1,14 @@
+export interface IDetailStationSearchDropdownProps {
+  filteredStationNames: string[];
+}
+
+export interface IDetailStationSearchInputProps {
+  searchInput: string;
+  stationNames: string[];
+  setSearchInput: (state: string) => void;
+  setFilteredStationNames: (state: string[]) => void;
+}
+
 export interface IDetailStationLineButtonProps {
   value: string;
   line: string;
